Patch by Andrew Sutherland:
On Debian/Ubuntu distributions, the standard for files in /usr/lib/debug does
not involve appending .debug to the filename.  It appears that the Fedora/RH
standard is to append .debug, which is what is currently implemented in
chronicle-query.  This patch rather simply adds an additional attempt to open
the debian-style standard before giving up.  The additional check should not
be dangerous, as Fedora/RH packages don't seem to install any files that would
match the second check attempt.

The documentation for the Debian policy is found here:
http://www.debian.org/doc/developers-reference/ch-best-pkging-practices.en.html#s-bpp-dbg